+++ app/models/brand.rb
@@ -75,4 +75,10 @@ class Brand < ApplicationRecord
     country = ISO3166::Country[country_code]
     country.translations[I18n.locale.to_s] || country.common_name || country.iso_short_name
   end
+
+  def formatted_description
+    return if description.blank?
+
+    Redcarpet::Markdown.new(Redcarpet::Render::HTML.new).render(description)
+  end
 end